renamed package and files to fit the established naming convention
