--- smacc2_sm_reference_library/simple_action_client_example/config/simple_action_client_example.yaml
@@ -1,7 +1,0 @@
-state_machine_rosparam_ws: /SimpleActionClientExample
-success_switch:
-   - type: state_reached
-     state_name: "simple_action_client_example::State2"
-failure_switch:
-   - type: timeout
-     duration: 10.0 # sec
