+++ smacc2_sm_reference_library/sm_simple_action_client/config/simple_action_client_example.yaml
@@ -1,0 +1,7 @@
+state_machine_rosparam_ws: /SmSimpleActionClient
+success_switch:
+   - type: state_reached
+     state_name: "sm_simple_action_client::State2"
+failure_switch:
+   - type: timeout
+     duration: 10.0 # sec
