+++ smacc2_sm_reference_library/sm_simple_action_client/config/simple_action_client_example_config.yaml
@@ -1,3 +1,3 @@
-simple_action_client_example:
+sm_simple_action_client:
   ros__parameters:
     signal_detector_loop_freq: 2000.0
