+++ smacc2_sm_reference_library/sm_simple_action_client/include/sm_simple_action_client/orthogonals/or_fibonacci.hpp
@@ -2,7 +2,7 @@
 #define ROBOT__STATE_MACHINE__ORTHOGONALS__OR_FIBONACCI_HPP_
 
 #include "smacc2/smacc_orthogonal.hpp"
-#include "simple_action_client_example/fibonacci_action_client/cl_fibonacci.hpp"
+#include "sm_simple_action_client/fibonacci_action_client/cl_fibonacci.hpp"
 
 namespace robot_state_machine
 {
