+++ smacc2_sm_reference_library/sm_simple_action_client/include/sm_simple_action_client/orthogonals/or_mode_select.hpp
@@ -2,7 +2,7 @@
 #define ROBOT_STATE_MACHINE__ORTHOGONALS__OR_MODE_SELECT_HPP_
 
 #include <smacc2/smacc_orthogonal.hpp>
-#include <simple_action_client_example/mode_selection_client/cl_mode_select.hpp>
+#include <sm_simple_action_client/mode_selection_client/cl_mode_select.hpp>
 
 namespace robot_state_machine
 {
