+++ smacc2_sm_reference_library/sm_simple_action_client/launch/sm_simple_action_client.launch
@@ -1,0 +1,5 @@
+<launch>
+  <node pkg="sm_simple_action_client" exec="sm_simple_action_client_node" name="sm_simple_action_client" output="screen">
+    <param from="$(find-pkg-share sm_simple_action_client)/config/sm_simple_action_client_config.yaml" />
+  </node>
+</launch>
