+++ smacc2_sm_reference_library/sm_simple_action_client/package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>simple_action_client_example</name>
+  <name>sm_simple_action_client</name>
   <version>2.3.15</version>
-  <description>The simple_action_client_example package</description>
+  <description>The sm_simple_action_client package</description>
   <maintainer email="smacc2@smacc2.cool">SMACC2 is cool</maintainer>
   <license>Apache-2.0</license>
   <license>MyLicense</license>
